feat: LSP autocomplete constructor fields (#5732)
# Description

## Problem

Part of https://github.com/noir-lang/noir/issues/1577

## Summary


![lsp-complete-constructor-fields](https://github.com/user-attachments/assets/7b822fd8-1852-492f-9d7f-48ae45e171a9)

^ The above is just an example of it working, I'm not encouraging using
constructors of std types 😄

## Additional Context

This is the last autocompletion I had in mind while implementing
autocompletion, so I'll likely won't send any other LSP PRs for a
while...

That said, if you think of other autocompletions (that don't involve
auto-importing, because that's a big thing) let me know (or capture an
issue) and we can work on it.

## Documentation

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.