chore: apply some new lints across workspace (#5736)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

This PR applies a couple of lints across the workspace, most notably
amongst these being that we disallow using the qualified name for an
import when we've imported that name into the current module already.

New lints:
```
trivial_casts = "warn"
trivial_numeric_casts = "warn"
unused_import_braces = "warn"
unused_qualifications = "warn"
```

I haven't applied these lints to `noirc_frontend` as it's got the most
violations and so we can apply that in a later PR.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [ ] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.