fix: Allow comptime code to use break without also being `unconstrained` (#5744)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

Allows `comptime` code to use `break` and `continue` without also
marking the functions `unconstrained`. This is more important with the
`unsafe { }` change since we don't want to have to wrap comptime code in
unsafe blocks, nor do we want to mark unconstrained functions used in a
comptime block as requiring `unsafe { }`.

## Additional Context



## Documentation\*

Check one:
- [ ] No documentation needed.
- [ ] Documentation included in this PR.
- [x] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.