feat: LSP auto-import completion (#5741)
# Description

## Problem

Part of https://github.com/noir-lang/noir/issues/1577

## Summary

I said I wasn't going to send any more LSP features, but I had this one
in the back of my head since I started working on autocompletion. I had
the idea of how to do it in my head, but I wanted to first work on the
easier stuff (regular autocompletion). It turns out auto-import wasn't
that hard to implement after all!


![lsp-autoimport](https://github.com/user-attachments/assets/9d2268fb-5caf-4b42-9bb3-b01f6ca40a9b)


![lsp-autoimport-nested](https://github.com/user-attachments/assets/ac4cd18a-d87d-4311-82d0-61f6d5a8dd19)

## Additional Context

In this PR every new imported item is added as a full path. Ideally,
like in Rust Analyzer, we'd group common use prefixes, add to existing
ones, etc. We can do that! But maybe it could be part of a follow-up PR,
or maybe we could do it much later. Or maybe `nargo fmt` could handle
this, so the auto-import doesn't do it, but once you save the file it's
done. That's why I didn't spend time on that in this PR.

There's another thing I noticed while working on this PR: Rust Analyzer
will offer matches that match **any** part of the name, not just the
beginning. So, in Noir, if you type `merkle` it should probably offer
`compute_merkle_root` as an auto-import completion, which is nice
because maybe you are looking for stuff related to "merkle" but don't
know the full name... but I didn't want to introduce that change in this
PR (it also works for every other autocompletion). But, as always, this
could be done later on.

## Documentation

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.