feat: (LSP) suggest names that match any part of the current prefix (#5752)
# Description

## Problem

Part of https://github.com/noir-lang/noir/issues/1577

## Summary

As I mentioned in #5741, I noticed Rust Analyzer will suggests methods
and names where the prefix you are writing matches any part of the name.
Well, kind of, the logic is not exactly that, but that's what's
implemented in this PR. It should make finding things much easier (if
you are looking for stuff related to "merkle" or "hasher" you just need
to type that).


![lsp-complete-any-part](https://github.com/user-attachments/assets/a65adc20-fc96-4682-b1c3-8961c434a133)

It works in any context, for example struct fields:


![lsp-complete-any-part-struct-field](https://github.com/user-attachments/assets/f52193ef-adf7-493b-afa5-dbae9009857e)

## Additional Context

None.

## Documentation

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.