refactor: simplify sequencer and l1 communication (#7989)
Fixes #7821. 

Looking to combine some of the components where there are interfaces
that don't really seem to be useful as there would likely be only one
implementation. When makes sense, I have joined the "wrapper" and the
thing it wrapped because extra wrapping is just confusing.

Starting with changing the global variables builder, since that was low
hanging. When looking at the `viem-tx-sender` it is more tedious as the
tests are heavily abusing that had two separate components.

Getting rid of the `gasEstimate` steps as viem seems to perform that
estimate step if no gas amount is provided, so might as well make it
less verbose, https://viem.sh/docs/contract/writeContract#gas-optional.

- Removes the `viem-reader`
- Removes the `viem-tx-sender`
- Removes the `receiver`
- Alter the `GlobalVariableBuilder` to directly read values
- Alter the `L1Publisher` to interact directly with the rollup and
availability oracle.
- Updates the tests