feat: fault-tolerant parsing of `fn` and `impl` (#5753)
# Description

## Problem

Part of https://github.com/noir-lang/noir/issues/1577

## Summary

Autocompletion wasn't triggering in function parameter types, or in impl
names. The reason is that when you start writing them the body will
still be missing, and that discarded the entire fn or impl. In this PR
we make that parse well, but produce an error (and create a fake empty
body).

Now autocompletion works fine in these scenarios:


![lsp-recover-fn](https://github.com/user-attachments/assets/e2358d49-c2e3-473b-bd4c-bfc6c0ff43e9)


![lsp-recover-impl](https://github.com/user-attachments/assets/a8fb1d44-86ce-4514-b167-f1ff6eb4b9b7)

## Additional Context

I'm not sure about the many new "parser error" introduced, maybe there
should be a generic "expected these tokens"? I think chumsky has one of
those errors but it's kind of tied to the parsing logic. On the other
hand adding new errors is simple, so maybe it's fine.

## Documentation

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.