feat: add `Expr::as_array`, `Expr::as_repeated_element_array` and same for slice (#5750)
# Description

## Problem

Part of #5668

## Summary

## Additional Context

I'm still not sure about the "repeated" name, and also about the
similarity of the methods between array and slice (maybe we can
distinguish array and slice with a boolean? not sure) but changing them
should be easy, so let me know.


## Documentation

Check one:
- [ ] No documentation needed.
- [ ] Documentation included in this PR.
- [x] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.