feat(perf): mem2reg function state for value loads to optimize across blocks (#5757)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

Part of general effort to reduce Brillig bytecode sizes. No issue as
found while working on other brillig optimizations.

## Summary\*

Now that we have a brillig gates diff I want to pull out individual
changes to live on their own to see only the individual impact across
all of our brillig tests.

This PR simply sees what is the last load we have performed on an
address across all blocks in a function. If we do not have a load on a
value, but any of the blocks still have store instructions to that
address, we can safely remove those store instructions.

## Additional Context



## Documentation\*

Check one:
- [X] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Tom French <15848336+TomAFrench@users.noreply.github.com>
Co-authored-by: jfecher <jake@aztecprotocol.com>