feat: user `super::` in LSP autocompletion if possible (#5751)
# Description

## Problem

Part of #1577

## Summary

We recently added `super::` to `use` so I thought it would be nice if
the auto-importer used that if possible.


![lsp-use-super](https://github.com/user-attachments/assets/4f8cbf37-f844-4293-832e-bc9685295837)

## Additional Context

This PR also includes a small refactor in the hover code because the
code to format a module path exists there too (but the two are different
enough that I didn't reuse them).

## Documentation

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: Tom French <15848336+TomAFrench@users.noreply.github.com>