feat: PG recursive verifier constructors based on stdlib inputs (#8052)
Update the PG recursive verifier to take stdlib
accumulators/verification_keys and stdlib proofs as opposed to their
native counterparts. This work will allow for a connection to be made
between the proofs/vkeys used in noir protocol circuits and those used
in the backend. The workflow will be roughly 1) construct stdlib
vkey/proof from their known native counterparts, 2) use `assert_equal`
to connect the witnesses of these objects to those received in the
corresponding acir `RecursionConstraint`, and 3) run the recursive
verifier as usual.