feat: Add `Expr::as_block` and `Expr::has_semicolon` (#5784)
# Description

## Problem

Part of #5668

## Summary

Both ExpressionKind and StatementKind will be exposed as `Expr` (as
discussed with Jake). So `Expr::as_block` will return the block
statements as a list of expressions. There's also the thing that some of
these statements might be `Semi` (well, all except maybe the last one)
and having to explicitly unwrap that with `as_semi` might be tedious or
bug-prone, so instead that's unwrapped automatically when you call any
of the `as_...` methods, but then there's an `has_semicolon` on `Expr`
to know if it has a semicolon.

## Additional Context

There's just those two methods in this PR because I wanted to validate
this approach. Next I'll add more `as_...` methods but for
StatementKind.

## Documentation

Check one:
- [ ] No documentation needed.
- [ ] Documentation included in this PR.
- [x] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.