fix: honor function visibility in LSP completion (#5809)
# Description

## Problem

Resolves #5808

## Summary

Reuses the same logic done when resolving a `use` or `Path` in the
compiler.

## Additional Context

It's debatable whether this functionality is desired, because sometimes
you add a function, want to use it and autocompletion doesn't suggest
it, and it takes a while for you to realize that you forgot to add
`pub`. However, a lot of the suggested functions come from the std or
external crates, and it doesn't make sense to reveal these private
functions. This is also how Rust Analyzer works too.

## Documentation

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.