fix(acir_gen): Nested dynamic array initialization (#5810)
# Description

## Problem\*

Resolves #5782 

## Summary\*

We are using our recursive `initialize_array_inner` function to also
check whether the outer AcirValue we are trying to initialize is a
dynamic array. We should be able to initialize a new AcirValue::Array
with internal AcirValue::DynamicArray types.

## Additional Context



## Documentation\*

Check one:
- [X] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.