feat: add `UnresolvedType::is_field` and `Expr::as_assign` (#5804)
# Description

## Problem

Part of #5668

## Summary

Also starts wrapping `LValue` too as an `Expr`, and makes
`Expr::as_member_access` work for the LValue that looks like a member
access.

## Additional Context

It seems the last commit of the previous PR didn't go out (sometimes git
will do that to me, not sure why) so it's included here.

## Documentation

Check one:
- [ ] No documentation needed.
- [ ] Documentation included in this PR.
- [x] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.