fix: bit shifting type checking (#5824)
# Description

## Problem\*

Resolves https://github.com/noir-lang/noir/issues/5823

## Summary\*

We check `rhs` differently than `lhs` in left and right bit shifts in a
series of hacky checks.
In the case of type variables, `lhs` and `rhs` may be swapped causing
the unification to be placed on lhs instead of rhs. I've reordered the
cases to prevent this.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.