feat!: wallet authwit management (#8128)
Added the ability of authorizing actions to be performed on behalf of
our accounts, both in public and private. Also revamped tests so they're
easier to follow and read their results.

## New commands

For private authwits

- `create-authwit [options] <functionName> <caller>`: Creates an
authorization witness that can be privately sent to a caller so they can
perform an action on behalf of the provided account.
- `add-authwit [options] <authwit> <authorizer>`: Adds an authorization
witness to the provided account, granting PXE access to the notes of the
authorizer so that it can be verified

Test showcasing usage
`./yarn-project/cli-wallet/test/flows/private_authwit_transfer.sh`

For public authwits

- `authorize-action [options] <functionName> <caller>`: Authorizes a
public call on the caller, so they can perform an action on behalf of
the provided account

Test showcasing usage
`./yarn-project/cli-wallet/test/flows/public_authwit_transfer.sh`

## Breaking changes

- `add-note` command now uses `-t, --transaction-hash` instead of `-h`
to not confuse it with the `--help` option. (thanks for the heads up,
@signorecello!)

## Other changes

- `bridge-fee-juice` now allows a `--no-wait` option to avoid polling
the network for the L2 messages to appear, and alternatively an
`--interval <seconds>` option to set polling interval (60s by default).

---------

Co-authored-by: esau <152162806+sklppy88@users.noreply.github.com>