chore(perf): Update to stdlib keccak for reduced Brillig code size (#5827)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

We can reduce the size of our keccak stdlib method in Brillig. There are
operations that are repeated across multiple places, different
operations we can perform depending on whether we are in an
unconstrained or constrained runtime, and we had an extra unnecessary
loop for building our `sliced_buffer` variable.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.