fix!: Check unused generics are bound (#5840)
# Description

## Problem\*

## Summary\*

Ran into this while working on arithmetic generics. It's possible to
have cases like:

```rs
fn foo<let N: u32>() {}

fn main() {
    foo();
}
```

For which `N` is unused but we don't issue a "type annotations needed"
error for it currently because it isn't present in the instantiated
function type of `foo`. This PR fixes this and issues the error.

## Additional Context

## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.