chore: don't require empty `Prover.toml` for programs with zero arguments but a return value (#5845)
…ents but a return value

# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

This is a small QOL change but we don't need to have a `Prover.toml`
file in the case where the circuit has no inputs but potential return
values. In this case we can tolerate a missing `Prover.toml` file.

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.