feat: simplify constant calls to `poseidon2_permutation`, `schnorr_verify` and `embedded_curve_add` (#5140)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

Now that we have rust implementations of all blackbox functions, we can
perform any pedersen operations with constant inputs at compile-time. We
couldn't do this before as it would have required an async
initialisation step for the wasm compiler but that is no longer an
issue.

I've done this by using compile-time flags to select the blackbox solver
we're using. If nargo is compiled with a non-bn254 field then it will
not perform these optimizations. The ultimate plan is for this solver to
be specified by the end user in a similar way to the way the field
definition is.





## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.