feat: warn on unused imports (#5847)
# Description

## Problem

Resolves #4762

## Summary

The compiler now produces a warning on unused imports.

## Additional Context

I don't know if this is the correct approach. I added a new HashSet to
track imported names, then those are removed as we import things. I
thought about tracking this in `ItemScope` but it's tricky because the
way things are they include imported and self things.

There's another thing: eventually it would be nice to warn on unused
types, like in Rust. For that maybe it would make sense to track this in
`ItemScope`... but given that we currently don't have visibility
modifiers for types, it can't be done right now. So maybe doing it just
for imports with specific code is fine.

I had to refactor a bit `resolve_path` because some paths were looked up
not using that method and so some things weren't marked as used... but
now `StandardPathResolver::new` is used in exactly one place in the
compiler.

## Documentation

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.