feat: remove unnecessary copying of vector size during reversal (#5852)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

## Summary\*

In order to calculate `vector_size - 1 - iterator`, rather than
initialising a register to `vector_size` and then decrementing it on
each loop iteration, we're currently recalculating it from scratch on
each iteration. I've modified this so that we can avoid these extra ops

## Additional Context



## Documentation\*

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.