fix(sha256): Fix upper bound when building msg block and delay final block compression under certain cases  (#5838)
# Description

## Problem\*

Resolves #5836 

## Summary\*

We accept a start index based upon the current block when parsing a
message. We should accurately base the upper bound to be based upon this
start index.

We also have special handling for building a message block but not
compressing it when the message is less than the block size. We need to
also do this handling for the last message block when we have a message
that is larger than the block size.

## Additional Context

~~sha256_var is currently getting warnings from the under constrained
check. It looks to only be happening on the new regression test added as
part of this PR that uses a larger message. The old sha256 tests do not
look to trigger these warnings which is strange. I am a bit unsure why I
am getting these warnings as msg block and msg block pointer are being
verified on each iteration of the loop.~~

<img width="965" alt="Screenshot 2024-08-27 at 12 04 58 PM"
src="https://github.com/user-attachments/assets/c71bff9d-bfea-4765-a134-91eca92c7806">

EDIT: This was only happening as my test was hashing constant values,
thus it was a dumb circuit. e.g the following:
```rust
fn main(result: pub [u8; 32]) {
    let headers = [102, 114, 111, 109, 58, 114, 117, 110, 110, 105, 101, 114, 46, 108, 101, 97, 103, 117, 101, 115, 46, 48, 106, 64, 105, 99, 108, 111, 117, 100, 46, 99, 111, 109, 13, 10, 99, 111, 110, 116, 101, 110, 116, 45, 116, 121, 112, 101, 58, 116, 101, 120, 116, 47, 112, 108, 97, 105, 110, 59, 32, 99, 104, 97, 114, 115, 101, 116];
    let hash = std::hash::sha256_var(headers, headers.len() as u64);
    assert_eq(hash, result);
}
```
The message needs to come from the inputs and the under-constrained
warnings go away.

## Documentation\*

Check one:
- [X] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.