feat: add `FunctionDef::body` (#5825)
# Description

## Problem

Part of #5668

## Summary

Also allows `quote { ... }.as_expr()` to work with statements and
L-values. Then adds an example that injects a `_context: Context`
parameter to functions annotated with `#[aztec]`, and to calls inside
those functions (any call for now). Also adds an incomplete `Expr::map`
to help doing this.

Oh, and allows unquoting `Value::Expr`, which wasn't possible before.

## Additional Context

A lot is remaining, but I thought this was a good cutting point. Next I
might work on #5828 because debugging comptime code without it is
tricky.

I didn't add docs for `Expr::map` yet as it's still not fully
functional. I'll add them once it works for in all cases.

## Documentation

Check one:
- [ ] No documentation needed.
- [ ] Documentation included in this PR.
- [x] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.