feat(meta): Comptime keccak (#5854)
# Description

## Problem\*

Resolves <!-- Link to GitHub Issue -->

Just under 1/3 of a public transfer in Aztec is due to a keccak hash of
the selector. We should ideally be simplifying this through our normal
compiler passes as the inputs are all known at compile time. This is
difficult at the moment due to having the same compilation pipeline for
SSA and ACIR and would require changing how we inline. For an easy win
we can enable comptime hashing of keccak.

This also just more generally provides another one of our foreign
functions in the comptime environment which will be needed in the
future.

## Summary\*

I added functions in the interpreter for `keccakf1600` and
`to_le_radix`. I also included a test under `compile_succes_empty` that
calls keccak in a comptime env.

The `assert_constant` builtin is also used by our byte decomposition
functions. For this we can just return true as we already check that no
non-comptime variables are referenced in comptime code.

## Additional Context



## Documentation\*

Check one:
- [] No documentation needed.
- [ ] Documentation included in this PR.
- [X] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: jfecher <jake@aztecprotocol.com>