chore: underconstrained check in parallel (#5848)
# Description

## Problem\*

underconstrained check can add significant time to the compilation
process

## Summary\*
perform the underconstrained check in parallel.
The PR parallelises the check per function. In the example I tried, it
gives no benefit as all the work is done in one function.
Then the PR also parallelise the 'merge sets' by doing it over chunks.
In the example I tried, it reduces the duration from 77s down to 1s.





## Additional Context



## Documentation\*

Check one:
- [X] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.

---------

Co-authored-by: TomAFrench <tom@tomfren.ch>
Co-authored-by: jfecher <jake@aztecprotocol.com>