feat: show backtrace on comptime assertion failures (#5842)
# Description

## Problem

Resolves #5828

## Summary

`CustomDiagnostic` didn't have a `call_stack` property because so far
this wasn't needed. I only added it for `FailingConstraint`, which is,
_I think_, the only case where a backtrace might be needed. In all other
cases (for example "wrong number of arguments" it might be clear why
it's failing without needing a backtrace).

Here's an example output:

```
error: Assertion failed
   ┌─ std/option.nr:33:16
   │
33 │         assert(self._is_some);
   │                -------------
   │
   ┌─ src/main.nr:8:9
   │
 8 │         foo();
   │         -----
   │
   ┌─ src/other.nr:2:5
   │
 2 │     bar();
   │     -----
   ·
 7 │     let _ = expr.as_integer().unwrap();
   │             --------------------------
   │

Aborting due to 1 previous error
```

## Additional Context

None.

## Documentation

Check one:
- [x] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist

- [x] I have tested the changes locally.
- [x] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.